+++ ImageObject.h
@@ -6,7 +6,7 @@ class ImageObject {
 	private:
 		int xt, yt, area, perimeter;
 	public:
-		ImageObject(int area, int perimeter, long int momentX, long int momentY);
+		ImageObject(int area, int perimeter, int momentX, int momentY);
 		const int getArea() const;
 		const int getPerimeter() const;
 		const int getXt() const;
